autorisation suppression

+++ view/frontend/postView.php
@@ -46,7 +46,12 @@
         ?>
             <p><strong><?= htmlspecialchars($comment['id_author']) ?></strong> le <?= $comment['creation_date'] ?></p>
             <p><?= nl2br(htmlspecialchars($comment['comment'])) ?></p>
-            <a href="index.php?action=deleteComment&id=<?= ($comment['id'])?>">Supprimer le commentaire</a>
+            
+            <?php if (!empty($_SESSION['id']) && ($_SESSION['role'] == 0 )) { ?>
+                <a href="index.php?action=deleteComment&id=<?= ($comment['id'])?>">Supprimer le commentaire</a>
+            <?php } ?>
+            
+            <a href="index.php?action=signalComment&id=<?= ($comment['id'])?>">Signaler le commentaire</a>
         <?php
         }
         ?>
